memory_agents/brand_reputation_monitor/requirements.txt
streamlit==1.50.0
memorisdk==2.3.0
langchain-openai==1.0.1
langchain-core==1.0.0
python-dotenv==1.1.1
brightdata-sdk==1.1.3
pydantic==2.12.3
openai==2.6.0